Conversion from CoNLL format
Lost non-terminal units are reconstructed, but any lost edges (from a node to its head) are assumed to have the C tag.
Can be fixed by an elaborate rule system or by machine learning, for example using a NN that learns the correct labels for a given UCCA DAG (see ucca-rnn).