Fix nonexistent head bug in conversion to CoNLL
If splitting to sentences, rarely a word has a dependency on a word in the next sentence. In that case, the head number will point to a word beyond the length of the sentence, yielding an invalid CoNLL file.
Now, such dependencies are removed, so if the word has other dependencies, they will be taken. If it has none, the last root will be used.