Prevent repeated edge and node actions
The parser frequently gets into loops where it keeps creating edges between the same nodes (which was not prevented up until now, unless it is the exact same edge with the same tag), or keeps creating unary parents (NODE) or implicit children (IMPLICIT).
This is now prevented by disallowing repeated edges (even if the tag is different), disallowing NODE when there are children, and disallowing IMPLICIT when there are no parents.
Running on the corpus verified that none of these disallowed actions is required in order to parse the corpus passages.