Combine valid_action_types into is_valid
It was never used on its own anyway, and now the code is more efficient as all the checks related to a specific action type are done in one place and only if this action type is currently being checked.
Also added some more restrictions on what constitutes a valid action, to prevent infinite loops in inference.
Also re-enabled the check that the oracle's action is valid (regardless of the Config().verify flag).
Also separated the indication that the state is a "finished" state to an attribute, because it is clearer than returning a boolean from transition().