Fix swap check by introducing swap index
Update the node's swap index according to the nodes before and after it.
Usually the swap index is usually just the index, and that is what it is initialized to.
If the buffer is not empty and the next node on it is not a terminal, it means that it is a non-terminal that was created at some point, probably before this node (because this method should be run just when this node is created). In that case, the buffer head's index will be lower than this node's index, and we will update this node's swap index to be the arithmetic average between the previous node (stack top) and the next node (buffer head). This will make sure that when we perform the validity check on the SWAP action, we will correctly identify this node as always having appearing before b (what is the current buffer head). Otherwise, we would prevent swapping this node with b even though it should be a legal action (because they have never been swapped before).