Assert node ratio limit when applying action
When it was in is_valid, the parser would just create the maximum number of non-terminals it could, and then continue doing other nonsense until the oracle happened to predict a NODE action.
This makes no sense, because have one node less than the limit is not a much better state than having the maximum number itself.
Therefore, if the parser is so stupid to just keep adding nodes, we want to let it do it and fail fast, just when it reaches the max.