Fix #9: enumerate features and keep weights matrix as array
During training the matrix is still a list of row arrays, so runtime suffers quite a lot, but this is necessary since we don't know all the features ahead of time.
Perhaps it will be more efficient to make it a 2D array from the beginning and vstack/append the rows each time.