Fix bug that caused "no action is available"
In predict_score, first the best-scoring action is attempted, and if invalid, all the other actions are sorted and the best-scoring valid action is chosen (assuming that as the parser gets better, usuallly the best-scoring one will be valid, and then we can do all this in O(n) time on average).
The problem is when there are two best-scoring actions (with the same score), in which case the action that we initially defined as "best" and the action that we take out of the list after sorting may be two different actions, resulting in the other best-scoring action being skipped.
Now we don't take out the best-scoring action after sorting. The runtime difference is negligible and the code is even simpler now.