Add dense perceptron class
Not used yet, but will allow using embedding features efficiently.
Also remove the "true labels" feature from the perceptron, as it was not used, because all of the original labels have always been in use (except for FINISH for some reason, but this should be investigated). It was originally added because the actions for all edge tags were added at construction, but now new actions are added on-the-fly and no edge actions exist at initialization.
 Renamed the original perceptron to sparse perceptron (which uses dictionaries from feature to weights vector).
 Added a Classifier super-class to handle common code. NeuralNetwork will also inherit from it.