Add criterion for MWE, but still commented out
This criterion is too broad (multiple terminal children or only center and function), because it catches a lot of cases with a copula (F) and a predicate (C).
A better criterion may be to also remove the cases where the F has a label of 'aux' in the dependency parse.
Will enable after dependency parsing is available.