Fix criteria for predicate nouns and adjectives
Now it doesn't matter if there's a Function inside.
In evaluation, there was a problem with moving common Functions to the root when it comes to predicate nouns/adjectives. For example, in "... Jackson is nuts ...", I would say "nuts" should be a predicate noun, but it wouldn't be identified as such because "nuts" itself does not have an incoming State edge; only "is nuts" does. But when we move the "is" to the root, suddenly "nuts" is left on its own and is identified. This changed the numbers a bit so the denominator in the recall is different for different parsers, which was unfortunate.
Previously predicate nouns could have only nouns; now they must have at least one noun and must not have adjectives.
Previously predicate adjectives could have only adjectives; now they must have at least one adjective and must not have nouns.